requirements.txt
Pillow==10.0.1
Pillow==10.3.0
streamlit==1.35.0
